sdm439-common: overlay: core: Allow more cached apps in the background
Google has provisioned this on the Pixel 5 (RQ1A.201205.010) to utilize
the 8 GB of RAM better, and I have also observed that a lightweight
selection of apps was hitting the cached app limit before exhausting RAM
on another device with 8 GB of RAM.

This shouldn't cause adverse effects on devices with less RAM (e.g. 4
GB) as the low memory killer should kick in long before this limit on
such devices.

Signed-off-by: DennySPb <dennyspb@gmail.com>
Change-Id: I902962c03b90a4432742ab8530edb5e4aff36999
Signed-off-by: rajkale99 <kaleraj.rk@gmail.com>
Signed-off-by: hsx02 <hsx02self@gmail.com>